+++ data/src/main/java/tachiyomi/data/chapter/ChapterRepositoryImpl.kt
@@ -37,6 +37,7 @@ class ChapterRepositoryImpl(
                         chapter.dateUpload,
                         chapter.version,
                         chapter.memo,
+                        chapter.pageCount,
                     )
                         .awaitAsOne()
                     chapter.copy(id = chapterId)
@@ -75,6 +76,7 @@ class ChapterRepositoryImpl(
                     version = chapterUpdate.version,
                     isSyncing = 0,
                     memo = chapterUpdate.memo?.let(MemoColumnAdapter::encode),
+                    pageCount = chapterUpdate.pageCount,
                 )
             }
         }
@@ -164,6 +166,7 @@ class ChapterRepositoryImpl(
         version: Long,
         isSyncing: Long,
         memo: JsonObject,
+        pageCount: Long,
     ): Chapter = Chapter(
         id = id,
         mangaId = mangaId,
@@ -180,5 +183,6 @@ class ChapterRepositoryImpl(
         lastModifiedAt = lastModifiedAt,
         version = version,
         memo = memo,
+        pageCount = pageCount,
     )
 }